projvar_blog/pyproject.toml
[tool.poetry]
name = "projvar-blog"
version = "0.1.0"
description = "Blog App"
authors = ["Amir <amir4vx@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
channels = "^4.1.0"
django-cors-headers = "^4.4.0"
django-filter = "^24.2"
django-model-utils = "^4.5.1"
django-split-settings = "^1.3.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.4.0"
psycopg2 = "^2.9.9"
pydantic = "^2.8.2"
pyyaml = "^6.0.1"
whitenoise = "^6.7.0"
gunicorn = "^22.0.0"
drf-yasg = "^1.21.7"
redis = "^5.0.7"
pika = "^1.3.2"
celery = {extras = ["redis"], version = "^5.4.0"}
django-redis = "^5.4.0"
django-celery = "3"
psycopg2-binary = "^2.9.9"
django-celery-results = "^2.5.1"


[tool.poetry.group.dev.dependencies]
colorlog = "^6.8.2"
django-debug-toolbar = "4.0.0"
pre-commit = "^3.7.1"
pytest = "^8.2.2"
pytest-django = "^4.8.0"
pytest-xdist = "^3.6.1"
model-bakery = "^1.18.2"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
